10_backend/requirements.txt
# FastAPI 핵심
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings>=2.0.0

# 데이터베이스 (PostgreSQL - AWS RDS)
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
asyncpg>=0.29.0  # 비동기 PostgreSQL 드라이버
greenlet>=3.0.0  # SQLAlchemy 비동기 지원

# 인증 & 보안
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4            # 비밀번호 해싱
python-multipart==0.0.6           # 폼 데이터 처리
cryptography>=41.0.7              # Fernet 암호화

# Rate Limiting
slowapi==0.1.9

# Email & Scheduling
apscheduler>=3.10.0              # Task scheduler
aiosmtplib>=3.0.0                # Async SMTP client
jinja2>=3.1.0                    # HTML template engine

# 기타
python-dotenv>=1.0.0              # 환경 변수
httpx>=0.25.2                     # HTTP 클라이언트

# ML & Data Processing
scikit-learn>=1.3.0
xgboost>=2.0.0                    # XGBoost 모델 (acc 73.47%)
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
